week8/w8d2/daily/website/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-phonenumber-field==7.0.1
phonenumberslite==8.13.2
sqlparse==0.4.3
tzdata==2022.7
